requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
django-filter==23.2
django-multitenant==3.2.1
django-rest-framework==0.1.0
django-rql==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lark-parser==0.11.0
lib-rql==2.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
packaging==23.1
pip-autoremove==0.10.0
psycopg2-binary==2.9.5
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.16.2
celery==5.2.7
redis==4.3.4
flower==0.9.0
django-celery-results==2.4.0
django-celery-beat==2.5.0
dj-database-url==2.1.0
gunicorn==21.2.0
django-cors-headers==4.2.0
